Modify amount field from Float to Integer.
Float are scary to handle.

Part of #54

+++ transport_nantes/stripe_app/models.py
@@ -40,7 +40,7 @@ class Donation(models.Model):
     mode = models.CharField(
         choices=[("SUB", "Subscription"), ("PAY", "One time donation")],
         max_length=50)
-    amount = models.FloatField()
+    amount = models.IntegerField()
     timestamp = models.DateTimeField(auto_now_add=True, null=True)
 
     def __str__(self):
